requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
redis==4.6.0
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.5.1
gunicorn==21.2.0
drf-spectacular==0.27.0
stripe>=14.0.0
django-cors-headers==4.3.0
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
python-decouple==3.8
django-rest-passwordreset>=1.0
Pillow==10.1.0
requests==2.31.0
django-redis==5.4.0